+++ api/tests/prescription/organization-learner/integration/domain/usecases/generate-organization-learners-username-and-temporary-password_test.js
@@ -2,9 +2,9 @@ import { NON_OIDC_IDENTITY_PROVIDERS } from '../../../../../../src/identity-acce
 import * as authenticationMethodRepository from '../../../../../../src/identity-access-management/infrastructure/repositories/authentication-method.repository.js';
 import { organizationLearnerIdentityRepository } from '../../../../../../src/identity-access-management/infrastructure/repositories/organization-learner-identity.repository.js';
 import * as userRepository from '../../../../../../src/identity-access-management/infrastructure/repositories/user.repository.js';
+import { OrganizationLearnerPasswordResetDTO } from '../../../../../../src/prescription/organization-learner/domain/models/OrganizationLearnerPasswordResetDTO.js';
 import { generateOrganizationLearnersUsernameAndTemporaryPassword } from '../../../../../../src/prescription/organization-learner/domain/usecases/generate-organization-learners-username-and-temporary-password.js';
 import { UserNotAuthorizedToUpdatePasswordError } from '../../../../../../src/shared/domain/errors.js';
-import { OrganizationLearnerPasswordResetDTO } from '../../../../../../src/shared/domain/models/OrganizationLearnerPasswordResetDTO.js';
 import * as userReconciliationService from '../../../../../../src/shared/domain/services/user-reconciliation-service.js';
 import * as organizationRepository from '../../../../../../src/shared/infrastructure/repositories/organization-repository.js';
 import { catchErr, databaseBuilder, domainBuilder, expect, sinon } from '../../../../../test-helper.js';
