+++ api/tests/prescription/organization-learner/unit/domain/usecases/generate-reset-organization-learners-password-csv_test.js
@@ -1,5 +1,5 @@
+import { OrganizationLearnerPasswordResetDTO } from '../../../../../../src/prescription/organization-learner/domain/models/OrganizationLearnerPasswordResetDTO.js';
 import { generateResetOrganizationLearnersPasswordCsvContent } from '../../../../../../src/prescription/organization-learner/domain/usecases/generate-reset-organization-learners-password-cvs-content.js';
-import { OrganizationLearnerPasswordResetDTO } from '../../../../../../src/shared/domain/models/OrganizationLearnerPasswordResetDTO.js';
 import { expect, sinon } from '../../../../../test-helper.js';
 
 describe('Unit | UseCases | Generate reset organization learners csv', function () {
